わーにんぐなおしたよ

+++ project/src/character.cpp
@@ -291,8 +291,8 @@ int CCharacter::CalDamage(float SkillAtkMul)
 {// 攻撃力 * 
 
 	int CalDamage =
-		((m_attack.GetBase() + m_attack.GetAddItem() + m_attack.GetBuffItem()) *
-		(m_attack.GetMulBuff() + m_attack.GetMulItem() + SkillAtkMul));
+		(int)(((m_attack.GetBase() + m_attack.GetAddItem() + m_attack.GetBuffItem()) *
+		(m_attack.GetMulBuff() + m_attack.GetMulItem() + SkillAtkMul)));
 
 	return CalDamage;
 }
